requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.5.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.41.0
django-crispy-forms==1.12.0
dnspython==2.1.0
filelock==3.0.12
Flask==1.1.2
Flask-PyMongo==2.3.0
googlemaps==4.4.5
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.1
pipenv==2020.11.15
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pylint==2.7.2
pymongo==3.11.3
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.4
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
